refactor: moved app props configs in separate package
