+++ src/main/java/ru/dankoy/korvotoanki/config/appprops/AppPropertiesConfig.java
@@ -1,4 +1,4 @@
-package ru.dankoy.korvotoanki.config;
+package ru.dankoy.korvotoanki.config.appprops;
 
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
