+++ src/main/java/ru/dankoy/korvotoanki/config/appprops/GoogleTranslatorProperties.java
@@ -1,4 +1,4 @@
-package ru.dankoy.korvotoanki.config;
+package ru.dankoy.korvotoanki.config.appprops;
 
 public interface GoogleTranslatorProperties {
